230613 게시물 좋아요 기능(로그인한 사용자만 좋아요 가능(취소도 가능), 수정 및 삭제 작성자만 가능하게 수정

+++ src/main/java/idusw/springboot/boradmwlee/domain/Board.java
@@ -31,5 +31,5 @@ public class Board { // 5 fields
     private LocalDateTime modDate;
 
     //like
-    private Integer boardLike;
+    private Long boardLike;
 }
